% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{pal_ndvi}
\alias{pal_ndvi}
\title{NDVI palette: gray to green}
\format{
colorRampPalette palette function gray to green
}
\description{
NDVI palette: gray to green, for raster
}
\keyword{datasets}
