% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{pal_ndvi2}
\alias{pal_ndvi2}
\title{NDVI palette: maroon to gold to green}
\format{
colorRampPalette palette function maroon to gold to green
}
\description{
NDVI palette: maroon to gold to green, for raster
}
\keyword{datasets}
