% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{pal_ndvi_n2}
\alias{pal_ndvi_n2}
\title{NDVI palette: maroon to gold to green}
\format{
colorNumeric palette function maroon to gold to green
}
\description{
NDVI palette: maroon to gold to green, scale -1 to 1 for legend.
}
\keyword{datasets}
