% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{park_meta}
\alias{park_meta}
\title{Meta data for parks}
\format{
dataframe with 1 row per park
}
\description{
A data frame of meta data for each of the parks, including total number of walks, walk starting coordinates, etc.
}
\keyword{datasets}
