% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_document.R
\docType{data}
\name{temp_pal}
\alias{temp_pal}
\title{Temperature palette: palette function}
\format{
palette function for Temperature
}
\description{
Temperature palette: palette function
}
\keyword{datasets}
